refactor: move map function inside lib/map
in order to use those function insvelte (without marionette)
